Refactor lib structure
